+++ src/backend/resolvers/mutations/organization/obm/index.js
@@ -3,6 +3,7 @@ const updateObmOrganization = require('./update/resolver')
 
 const obmServiceResolvers = require('./service')
 const obmServiceCategoryResolvers = require('./serviceCategory')
+const obmTypeResolvers = require('./type')
 const obmRelationalResolvers = require('./relational-resolvers')
 
 module.exports = {
@@ -10,5 +11,6 @@ module.exports = {
   updateObmOrganization,
   ...obmServiceResolvers,
   ...obmServiceCategoryResolvers,
+  ...obmTypeResolvers,
   ...obmRelationalResolvers,
 }