+++ src/backend/resolvers/mutations/organization/obm/relational-resolvers/index.js
@@ -1,13 +1,15 @@
 const serviceAndServiceCategoryResolvers = require('./serviceAndServiceCategory')
 const obmAndServiceResolvers = require('./obmAndService')
 const obmAndPersonResolvers = require('./obmAndPerson')
-const obmAndPayer = require('./obmAndPayer')
+const obmAndObmTypeResolvers = require('./obmAndObmType')
+const obmAndPayerResolvers = require('./obmAndPayer')
 const obmAndKeyEventResolvers = require('./obmAndKeyEvent')
 
 module.exports = {
   ...serviceAndServiceCategoryResolvers,
   ...obmAndServiceResolvers,
   ...obmAndPersonResolvers,
-  ...obmAndPayer,
+  ...obmAndObmTypeResolvers,
+  ...obmAndPayerResolvers,
   ...obmAndKeyEventResolvers,
 }