+++ src/backend/resolvers/mutations/organization/obm/type/create.js
@@ -1,0 +1,7 @@
+const createObmType = (parent, { input }, { pulseCoreDb }, info) =>
+  pulseCoreDb
+    .collection('obms.types')
+    .insertOne({ ...input })
+    .then(({ ops }) => ops[0])
+
+module.exports = createObmType