+++ src/backend/resolvers/queries/organizations/obm/index.js
@@ -1,6 +1,7 @@
 const obmOrganizations = require('./obm')
 const obmServices = require('./service')
 const obmServicesCategories = require('./serviceCategory')
+const obmTypes = require('./type')
 const obmKeyEvents = require('./keyEvents')
 const joins = require('./joins')
 const views = require('./views')
@@ -9,6 +10,7 @@ module.exports = {
   obmServices,
   obmOrganizations,
   obmServicesCategories,
+  obmTypes,
   obmKeyEvents,
   ...joins,
   ...views,