+++ src/backend/typeDefs/mutations/index.js
@@ -160,12 +160,19 @@ const mutationType = gql`
       input: DeleteObmServiceCategoryInput!
     ): DeleteObmServiceCategoryPayload!
 
+    createObmType(input: CreateObmTypeInput!): CreateObmTypePayload!
+    updateObmType(input: UpdateObmTypeInput!): UpdateObmTypePayload!
+    deleteObmType(input: DeleteObmTypeInput!): DeleteObmTypePayload!
+
     connectObmServiceAndObmServiceCategory(
       input: ConnectObmServiceAndObmServiceCategoryInput!
     ): ConnectObmServiceAndObmServiceCategoryPayload!
     connectObmAndObmService(
       input: [ConnectObmAndObmServiceInput!]!
     ): [ConnectObmAndObmServicePayload!]!
+    connectObmAndObmType(
+      input: ConnectObmAndObmTypeInput!
+    ): ConnectObmAndObmTypePayload!
     connectObmAndPerson(
       input: [ConnectObmAndPersonInput!]!
     ): [ConnectObmAndPersonPayload!]!