+++ src/backend/typeDefs/mutations/organization/obm/relational-typeDefs/index.js
@@ -1,5 +1,6 @@
 const serviceAndServiceCategoryTypeDefs = require('./serviceAndServiceCategory')
 const obmAndServiceTypeDefs = require('./obmAndService')
+const obmAndObmTypeTypeDefs = require('./obmAndObmType')
 const obmAndPersonTypeDefs = require('./obmAndPerson')
 const obmAndPayerTypeDefs = require('./obmAndPayer')
 const obmAndKeyEvent = require('./obmAndKeyEvent')
@@ -7,6 +8,7 @@ const obmAndKeyEvent = require('./obmAndKeyEvent')
 module.exports = [
   ...serviceAndServiceCategoryTypeDefs,
   ...obmAndServiceTypeDefs,
+  ...obmAndObmTypeTypeDefs,
   ...obmAndPersonTypeDefs,
   ...obmAndPayerTypeDefs,
   ...obmAndKeyEvent,