+++ src/backend/typeDefs/mutations/organization/obm/relational-typeDefs/obmAndObmType/index.js
@@ -1,0 +1,3 @@
+const connectObmAndObmTypeTypeDefs = require('./upsert')
+
+module.exports = [connectObmAndObmTypeTypeDefs]